+++ src/components/_common/cardV2/MiniCardV2.module.scss
