+++ src/components/parts/product-card-mini/ProductCardMiniPart.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { usePageContentProps } from 'store/pageContext';
 import { CardTargetProps, getCardProps } from 'components/_common/card/card-utils';
-import { CardV2 } from 'components/_common/cardV2/CardV2';
+import { CardV2 } from 'components/_common/cardV2/MiniCardV2';
 import { EditorHelp } from 'components/_editor-only/editor-help/EditorHelp';
 import { PartComponentProps, PartType } from 'types/component-props/parts';
